+++ packages/core/src/constants.ts
@@ -1,31 +1,34 @@
 import type { ImgFormatType, PluginOptions } from './types'
 import { AnyObject } from './types'
 
-export const IMG_FORMATS_ENUM: { [key: string]: ImgFormatType } = {
+export const IMG_FORMATS_ENUM = {
   png: 'png',
   jpg: 'jpg',
   jpeg: 'jpeg',
   webp: 'webp',
   avif: 'avif',
   tiff: 'tiff',
-  gif: 'gif'
+  gif: 'gif',
+  svg: 'svg'
 } as const
-
-const imgFormats: ImgFormatType[] = []
-for (const key in IMG_FORMATS_ENUM) {
-  if (Object.prototype.hasOwnProperty.call(IMG_FORMATS_ENUM, key)) {
-    imgFormats.push(IMG_FORMATS_ENUM[key])
-  }
-}
 
 export const DEFAULT_CONFIG: PluginOptions = {
   quality: 80,
   enableDev: false,
   enableDevWebp: false,
   enableWebp: false,
-  include: imgFormats,
+  includes: /\.(png|jpe?g|gif|webp|svg|avif)(\?.*)?$/i,
+  excludes: '',
   cacheDir: 'node_modules/.cache/vite-plugin-image',
   sharpConfig: {},
+  svgoConfig: {
+    plugins: [
+      'preset-default',
+      { name: 'removerXMLNS', fn: () => null},
+      { name: 'removeViewBox' }
+    ],
+    js2svg: { indent: 2, pretty: true }
+  },
   compatibility: false,
   bodyWebpClassName: 'webp',
   filter: () => true