+++ packages/core/src/index.ts
@@ -41,25 +41,25 @@ export default function ImageTools(
       }
       server.middlewares.use(async (req, res, next) => {
         const url = req.url || ''
-        if (!filterImage(url)) return next()
-
+        const isFilter = await filterImage(url)
+        if (!isFilter) return next()
         try {
           const filePath = decodeURIComponent(
             path.resolve(process.cwd(), url.split('?')[0].slice(1) || '')
           )
-
-          const isTrue = await handleFilterPath(url)
-          if (!isTrue) {
-            return next()
-          }
-
+      
+          // const isTrue = await handleFilterPath(url)
+          // if (!isTrue) {
+          //   return next()
+          // }
+      
           const { ext } = parse(filePath)
           const type = enableDevWebp
             ? IMG_FORMATS_ENUM.webp
             : ext.replace('.', '')
 
           const buffer = await processImage(filePath)
-
+          
           if (!buffer) {
             return next()
           }